services/api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic[email]>=2.6.0
python-dotenv>=1.0.0
httpx>=0.26.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
google-generativeai>=0.3.2
supabase>=2.3.0
prisma>=0.11.0
aiohttp>=3.9.0
Pillow>=10.0.0
